ci: restore the CI suite to green and cut its runner-minute cost (Task 174)
Every workflow has been red since 2026-07-04. There were two independent
causes, one of them not in this repository.

Infrastructure (not fixable here). Every job of every workflow completes in
~2 seconds with no runner assigned and zero steps executed — GitHub declining
to *schedule* the jobs, not the jobs failing. It persisted across the monthly
billing reset, so it is not simple minute exhaustion but an account-level block
on a metered private repository. Only the owner can clear it (make the
repository public, raise the spending limit, or use self-hosted runners); the
signature and the options are written up in docs/development/ci.md.

Real gate failures, hidden behind that block. With CI unable to run for ~87
commits, two required gates rotted:

  * Static analysis: 33 findings (11 gofmt, 7 staticcheck, 6 unparam, 4
    errcheck, 3 nilerr, 1 ineffassign, 1 unconvert). Fixed at the source rather
    than suppressed, except where the code was already right — those carry a
    //nolint with the reason, matching the convention already in the tree.

    The one behaviour change is in handlers.consumeHSMAuditLogs, flagged as a
    deprecated call: it acknowledged the YubiHSM device audit log *before*
    persisting it, so a store failure destroyed the only copy of records
    nothing can reconstruct. It now fetches, persists, and only then
    acknowledges, as internal/hsmaudit's collector already did. Re-fetching an
    unacknowledged segment is safe because StoreHSMAuditEntries
    inserts-or-ignores on the entry number.

  * Coverage ratchet: 6 packages had drifted below the baseline, the largest
    being internal/hsm (-12.6pp) whose Task 171 native driver is hardware-gated
    and unreachable from HSM-free CI. The baseline was also stale in two ways:
    it was missing the 11 packages added since Task 119, and it carried
    internal/cliout at 95.2% — a package that exists only in an uncommitted
    working tree, so the committed baseline had been generated from a dirty
    checkout. Regenerated from a clean tree; the total is unchanged at 62.4%
    and the diff is mostly gains (grpcapi +19.0pp, pki +9.4pp, acme +5.1pp).

Runner-minute cost. The suite as configured cannot fit a private repository's
allowance — on 2026-07-02 the Test workflow alone ran 52 times in one day.
Two changes, both preserving coverage:

  * The four advisory (continue-on-error) suites in enterprise-ci.yaml — chaos,
    -race, interop and the benchmark diff — now run on a nightly schedule and
    on manual dispatch instead of on every push. They gate nothing by
    construction and were the most expensive half of the workflow.
  * test.yaml, the upstream project's original job, was still matching
    branches: ['**'] with no concurrency group, so every push on every branch
    started a KeyCloak container and a SoftHSM token and ran to completion even
    when immediately superseded. Scoped to enterprise with cancel-in-progress.

Verified locally against SoftHSM, PostgreSQL 16 and a strict docs build: the
integration suite, fuzz smoke, FIPS (build + unit + HSM smoke), DR store
integrity, docs structure, OpenAPI drift, static analysis, coverage ratchet,
govulncheck, SBOM and the MkDocs site all pass. The four advisory suites and
the kind-based k8s smoke were not run locally.

+++ server/internal/hsmaudit/verify.go
@@ -322,7 +322,7 @@ func VerifyChainFromGenesis(entries []hsm.AuditLogEntry, unlogged Unlogged) *Seg
 // tolerating the wrap. Half the counter space is treated as "ahead" and the
 // other half as "behind", the standard serial-number comparison.
 func isForward(want, got uint16) bool {
-	return uint16(got-want) < 0x8000
+	return got-want < 0x8000
 }
 
 // gapSize returns how many entry numbers were skipped between want and got.
